+++ BEmu_cpp/headers/IntradayTickRequest/ElementIntradayTickData.h
@@ -39,8 +39,8 @@ namespace BEmu
 				~ElementIntradayTickData();
 
 				virtual Name name() const;
-				virtual int numValues() const { return 1; }
-				virtual int numElements() const { return 0; }
+				virtual size_t numValues() const { return 1; }
+				virtual size_t numElements() const { return 0; }
 		
 				virtual bool isArray() const { return false; }
 				virtual bool isComplexType() const { return false; }
